Change timesync function to a callback function.
