+++ src/bacnet/basic/service/h_ts.h
@@ -39,6 +39,11 @@
 #include "bacnet/datetime.h"
 #include "bacnet/wp.h"
 
+typedef void (*handler_timesync_set_callback_t)(
+    BACNET_DATE *bdate,
+    BACNET_TIME *btime,
+    bool utc);
+
 #ifdef __cplusplus
 extern "C" {
 #endif /* __cplusplus */
@@ -81,7 +86,9 @@ extern "C" {
     bool handler_timesync_recipient_address_set(
         unsigned index,
         BACNET_ADDRESS * address);
-
+    BACNET_STACK_EXPORT
+    void handler_timesync_set_callback_set(
+        handler_timesync_set_callback_t cb);
 #ifdef __cplusplus
 }
 #endif /* __cplusplus */
