+++ src/bacnet/datetime.h
@@ -305,12 +305,6 @@ bool datetime_local(BACNET_DATE *bdate,
     int16_t *utc_offset_minutes,
     bool *dst_active);
 BACNET_STACK_EXPORT
-bool datetime_timesync(
-    BACNET_DATE *bdate,
-    BACNET_TIME *btime,
-    bool utc);
-
-BACNET_STACK_EXPORT
 void datetime_init(void);
 
 #ifdef __cplusplus
